requirements.txt
pdfplumber==0.11.4
pypdf==5.4.0
pandas==2.2.2
openpyxl==3.1.5
PyYAML==6.0.2
pytesseract==0.3.13
pdf2image==1.17.0
pytest==8.3.5
Flask==3.0.3
